Update e3dsCore.min.js
e3dsCore.min.js suppresses the parent console.log, which means that if we call "console.log" on a parent page that uses this script, it will not be called. The code updated suggests a modification that suppresses console.log only in the script itself, without affecting the parent console.log.